requirements.txt
fastapi==0.79.0
python-dotenv==0.20.0
requests==2.28.1
shortuuid==1.0.9
twilio==7.12.1
